kuring-69 [수정] 교직원 검색 로직을 :data:staff로 옮김
