+++ app/src/main/java/com/ku_stacks/ku_ring/ui/main/search/SearchViewModel.kt
@@ -3,13 +3,13 @@ package com.ku_stacks.ku_ring.ui.main.search
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.ku_stacks.ku_ring.data.api.SearchClient
-import com.ku_stacks.ku_ring.data.mapper.toStaffList
 import com.ku_stacks.ku_ring.domain.Notice
 import com.ku_stacks.ku_ring.domain.Staff
 import com.ku_stacks.ku_ring.notice.api.NoticeClient
 import com.ku_stacks.ku_ring.notice.mapper.toNoticeList
 import com.ku_stacks.ku_ring.notice.repository.NoticeRepository
+import com.ku_stacks.ku_ring.staff.mapper.toStaffList
+import com.ku_stacks.ku_ring.staff.remote.StaffClient
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 import io.reactivex.rxjava3.schedulers.Schedulers
@@ -19,7 +19,7 @@ import javax.inject.Inject
 @HiltViewModel
 class SearchViewModel @Inject constructor(
     private val noticeRepository: NoticeRepository,
-    private val searchClient: SearchClient,
+    private val staffClient: StaffClient,
     private val noticeClient: NoticeClient,
 ) : ViewModel() {
 
@@ -40,7 +40,7 @@ class SearchViewModel @Inject constructor(
     fun searchStaff(keyword: String) {
         Timber.e("search staff $keyword")
         disposable.add(
-            searchClient.fetchStaffList(keyword)
+            staffClient.fetchStaffList(keyword)
                 .subscribeOn(Schedulers.io())
                 .filter { it.isSuccess }
                 .map { staffResponse -> staffResponse.toStaffList() }
