+++ data/staff/src/main/java/com/ku_stacks/ku_ring/staff/remote/StaffService.kt
@@ -1,11 +1,11 @@
-package com.ku_stacks.ku_ring.data.api
+package com.ku_stacks.ku_ring.staff.remote
 
-import com.ku_stacks.ku_ring.data.api.response.SearchStaffListResponse
+import com.ku_stacks.ku_ring.staff.remote.response.SearchStaffListResponse
 import io.reactivex.rxjava3.core.Single
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-interface SearchService {
+interface StaffService {
     @GET("v2/staffs/search")
     fun fetchStaffs(
         @Query("content") content: String,
