+++ data/staff/src/main/java/com/ku_stacks/ku_ring/staff/remote/response/SearchStaffListResponse.kt
@@ -1,4 +1,4 @@
-package com.ku_stacks.ku_ring.data.api.response
+package com.ku_stacks.ku_ring.staff.remote.response
 
 import com.google.gson.annotations.SerializedName
